Moved flowTest functions to a folder under cli
